feat: реализовать проверку удаления чужой карточки

+++ controllers/cards.js
@@ -5,6 +5,7 @@ const {
   OK,
   BAD_REQUEST,
   NOT_FOUND,
+  FORBIDDEN,
   INTERNAL_SERVER,
 } = require('../utils/errors');
 
@@ -33,16 +34,27 @@ module.exports.createCard = (req, res) => {
 };
 
 module.exports.deleteCard = (req, res) => {
-  Card.findByIdAndRemove(req.params.cardId).orFail()
-    .then((card) => {
-      res.status(OK).send({ message: `Карточка _id:${card._id} удалена` });
+  Card.findById(req.params.cardId).orFail()
+    .then((cardDoc) => {
+      console.log(req.user._id);
+      console.log(cardDoc.owner.toString());
+      if (req.user._id !== cardDoc.owner.toString()) {
+        const err = new Error('Нельзя удалять чужую карточку');
+        err.name = 'Forbidden';
+        return Promise.reject(err);
+      }
+      return Card.findByIdAndRemove(req.params.cardId).orFail();
     })
+    .then((card) => res.status(OK).send({ message: `Карточка _id:${card._id} удалена` }))
     .catch((err) => {
       if (err instanceof mongoose.Error.CastError) {
         return res.status(BAD_REQUEST).send({ message: 'Переданы некорректные данные карточки.' });
       }
       if (err instanceof mongoose.Error.DocumentNotFoundError) {
         return res.status(NOT_FOUND).send({ message: 'Не найдена карточка с указанным _id.' });
+      }
+      if (err.name === 'Forbidden') {
+        return res.status(FORBIDDEN).send({ message: err.message });
       }
       return res.status(INTERNAL_SERVER).send({ message: 'Произошла ошибка на сервере.' });
     });
